extract MuiPage interface to separate file
so that it can be imported without touching pagesApi.js
