+++ docs/src/pages.ts
@@ -1,43 +1,5 @@
 import pagesApi from './pagesApi';
-
-export interface MuiPage {
-  pathname: string;
-  children?: MuiPage[];
-  disableDrawer?: boolean;
-  icon?: string;
-  /**
-   * Indicates if the pages are regarding some legacy API.
-   */
-  legacy?: boolean;
-  /**
-   * Indicates if the pages are only available in some plan.
-   * @default 'community'
-   */
-  plan?: 'community' | 'pro' | 'premium';
-  /**
-   * In case the children have pathnames out of pathname value, use this field to scope other pathnames
-   */
-  scopePathnames?: string[];
-  /**
-   * Pages are considered to be ordered depth-first.
-   * If a page should be excluded from this order, set `order: false`.
-   * You want to set `inSideNav: false` if you don't want the page to appear in an ordered list e.g. for previous/next page navigation.
-   */
-  inSideNav?: boolean;
-  /**
-   * Props spread to the Link component
-   */
-  linkProps?: Record<string, unknown>;
-  subheader?: string;
-  /**
-   * Overrides the default page title.
-   */
-  title?: string;
-}
-
-export interface OrderedMuiPage extends MuiPage {
-  ordered?: true;
-}
+import type { MuiPage, OrderedMuiPage } from './MuiPage';
 
 const pages: readonly MuiPage[] = [
   {
@@ -368,4 +330,5 @@ const pages: readonly MuiPage[] = [
   { pathname: '/blog', title: 'Blog', icon: 'BookIcon' },
 ];
 
+export type { MuiPage, OrderedMuiPage };
 export default pages;
